revert large refactoring
