--- src/merkle_tree/bitcoin_script/mod.rs
@@ -1,2 +1,0 @@
-mod twin_proof;
-pub use twin_proof::*;
